Refactored Module to be in ...modules
