+++ src/org/team5940/robot_core/modules/interfaces/OwnableModule.java
@@ -1,5 +1,7 @@
 package org.team5940.robot_core.modules.interfaces;
 
+import org.team5940.robot_core.modules.Module;
+
 /**
  * This extension of Module provides the interface to lock access to a Module to a particular Thread. This is useful in preventing unintentional concurrent usage, particularly of actuators.
  * @author David Boles
